Added original python code to build a tune book
This is the code I wrote for building son and tune books. It's currently being used for the Paverty Bush Band song and tune books, and the Canberra English Session tunebook.
It uses abcm2ps to create SVG files that get included.
The resulting HTML can be converted to PDF by loading it with Chrome browser and printing to PDF.